refactor: network utils moved to common lib
