--- import-service/src/utils/response.ts
@@ -1,22 +1,0 @@
-import {
-    StatusCodes,
-} from 'http-status-codes';
-
-const addHeaderCORS = (response) => {
-    response.headers = {
-        ...response.headers,
-        'Access-Control-Allow-Origin': '*',
-        'Access-Control-Allow-Credentials': true,
-    }
-
-    return response;
-}
-
-export const toSuccess = (data) => addHeaderCORS({
-    statusCode: StatusCodes.OK,
-    body: data,
-});
-
-export const toAccepted = () => addHeaderCORS({
-   statusCode: StatusCodes.ACCEPTED,
-});
