+++ product-service/src/handlers/product-by-id.ts
@@ -2,7 +2,7 @@ import { APIGatewayProxyHandler } from 'aws-lambda';
 import 'source-map-support/register';
 
 import {getById} from '../dataStore'
-import {toNotFound, toServerError, toSuccess, toValidationError} from '../utils/response';
+import {toNotFound, toServerError, toSuccess, toValidationError} from '../../../common/src/utils/response';
 import {idSchema} from './schemas/id';
 
 export const getProductsById: APIGatewayProxyHandler = async (event) => {
